refactor(company): improve code structure
