+++ src/data/post.mock.ts
@@ -1,0 +1,42 @@
+import { Post } from "../types";
+import { PostApiRepository } from "./post";
+import { DataRepository } from "./repository";
+
+export const mockPosts: Post[] = [
+  {
+    userId: 1,
+    id: 8,
+    title: "dolorem dolore est ipsam",
+    body: "dignissimos aperiam dolorem qui eum\nfacilis quibusdam animi sint suscipit qui sint possimus cum\nquaerat magni maiores excepturi\nipsam ut commodi dolor voluptatum modi aut vitae",
+  },
+  {
+    userId: 1,
+    id: 9,
+    title: "nesciunt iure omnis dolorem tempora et accusantium",
+    body: "consectetur animi nesciunt iure dolore\nenim quia ad\nveniam autem ut quam aut nobis\net est aut quod aut provident voluptas autem voluptas",
+  },
+  {
+    userId: 2,
+    id: 11,
+    title: "et ea vero quia laudantium autem",
+    body: "delectus reiciendis molestiae occaecati non minima eveniet qui voluptatibus\naccusamus in eum beatae sit\nvel qui neque voluptates ut commodi qui incidunt\nut animi commodi",
+  },
+  {
+    userId: 2,
+    id: 12,
+    title: "in quibusdam tempore odit est dolorem",
+    body: "itaque id aut magnam\npraesentium quia et ea odit et ea voluptas et\nsapiente quia nihil amet occaecati quia id voluptatem\nincidunt ea est distinctio odio",
+  },
+];
+
+export class PostRepository
+  extends DataRepository<Post>
+  implements PostApiRepository
+{
+  public type = "Post" as const;
+
+  async load(): Promise<PostRepository> {
+    this.data = await Promise.resolve(mockPosts);
+    return this;
+  }
+}
