+++ src/routes/company.ts
@@ -1,10 +1,9 @@
-import { postService } from "../services";
-import { ChildService } from "../services/post";
+import { postService, ChildService } from "../services";
 import { ApiRequest, ApiResponse, Post, User } from "../types";
 import { ValidationError, handleExternalError } from "../utils";
 
 export function validateCompanyParameters(req?: ApiRequest) {
-  if (!req?.query?.company || typeof req?.query?.company !== "string") {
+  if (!req?.query?.companyName || typeof req?.query?.companyName !== "string") {
     throw new ValidationError(
       "Need something like ?companyName=Romaguera, " +
         "where companyName is the name of the company to fetch."
@@ -16,13 +15,16 @@ export function validateCompanyParameters(req?: ApiRequest) {
   return { companyName };
 }
 
+export type CompanyQuery = {
+  companyName: string;
+};
 export const companyRouteFactory = (
   postService: ChildService<User, Post>,
-  paramValidator: (req?: ApiRequest) => {
+  paramValidator: (req?: ApiRequest<CompanyQuery>) => {
     companyName: string;
   } = validateCompanyParameters
 ) =>
-  function companyRoute(req?: ApiRequest): Promise<ApiResponse> {
+  function companyRoute(req?: ApiRequest<CompanyQuery>): Promise<ApiResponse> {
     const { companyName } = paramValidator(req);
 
     return postService
