+++ src/routes/profile.spec.ts
@@ -1,83 +1,7 @@
 import { expect, it, describe, beforeEach, vi } from "vitest";
 
-import type { User, Post } from "../types";
-
-// Mock data for testing
-const mockUsers: User[] = [
-  {
-    id: 1,
-    name: "Leanne Graham",
-    username: "Bret",
-    email: "Sincere@april.biz",
-    address: {
-      street: "Kulas Light",
-      suite: "Apt. 556",
-      city: "Gwenborough",
-      zipcode: "92998-3874",
-      geo: {
-        lat: "-37.3159",
-        lng: "81.1496",
-      },
-    },
-    phone: "1-770-736-8031 x56442",
-    website: "hildegard.org",
-    company: {
-      name: "Romaguera-Crona",
-      catchPhrase: "Multi-layered client-server neural-net",
-      bs: "harness real-time e-markets",
-    },
-  },
-  {
-    id: 2,
-    name: "Ervin Howell",
-    username: "Antonette",
-    email: "Shanna@melissa.tv",
-    address: {
-      street: "Victor Plains",
-      suite: "Suite 879",
-      city: "Wisokyburgh",
-      zipcode: "90566-7771",
-      geo: {
-        lat: "-43.9509",
-        lng: "-34.4618",
-      },
-    },
-    phone: "010-692-6593 x09125",
-    website: "anastasia.net",
-    company: {
-      name: "Deckow-Crist",
-      catchPhrase: "Proactive didactic contingency",
-      bs: "synergize scalable supply-chains",
-    },
-  },
-];
-
-const mockPosts: Post[] = [
-  {
-    userId: 1,
-    id: 8,
-    title: "dolorem dolore est ipsam",
-    body: "dignissimos aperiam dolorem qui eum\nfacilis quibusdam animi sint suscipit qui sint possimus cum\nquaerat magni maiores excepturi\nipsam ut commodi dolor voluptatum modi aut vitae",
-  },
-  {
-    userId: 1,
-    id: 9,
-    title: "nesciunt iure omnis dolorem tempora et accusantium",
-    body: "consectetur animi nesciunt iure dolore\nenim quia ad\nveniam autem ut quam aut nobis\net est aut quod aut provident voluptas autem voluptas",
-  },
-  {
-    userId: 2,
-    id: 11,
-    title: "et ea vero quia laudantium autem",
-    body: "delectus reiciendis molestiae occaecati non minima eveniet qui voluptatibus\naccusamus in eum beatae sit\nvel qui neque voluptates ut commodi qui incidunt\nut animi commodi",
-  },
-  {
-    userId: 2,
-    id: 12,
-    title: "in quibusdam tempore odit est dolorem",
-    body: "itaque id aut magnam\npraesentium quia et ea odit et ea voluptas et\nsapiente quia nihil amet occaecati quia id voluptatem\nincidunt ea est distinctio odio",
-  },
-];
+import { mockPosts } from "../data/post.mock";
+import { mockUsers } from "../data/user.mock";
 
 // Mock implementation of fetchData for testing
 const fetchDataMock = <T>(url: string): Promise<T> => {
