+++ src/services/index.ts
@@ -1,12 +1,10 @@
 import { PostRepository, UserRepository } from "../data";
-import { Post, User } from "../types";
-import { ChildDataService } from "./post";
+import { PostService } from "./post";
 
 const userRepo = new UserRepository();
 const postRepo = new PostRepository();
 
-export const postService = new ChildDataService<User, Post>(
-  userRepo,
-  postRepo,
-  "userId"
-);
+export const postService = new PostService(userRepo, postRepo);
+
+export * from "./child";
+export * from "./post";
